Move get_networks to views
