+++ raiden/transfer/views.py
@@ -33,6 +33,7 @@ from raiden.utils.typing import (
     Set,
     TokenAddress,
     TokenNetworkAddress,
+    Tuple,
     Union,
 )
 
@@ -592,3 +593,24 @@ def detect_balance_proof_change(
                     if our_state_updated:
                         assert current_channel.our_state.balance_proof, MYPY_ANNOTATION
                         yield current_channel.our_state.balance_proof
+
+
+def get_networks(
+    chain_state: ChainState,
+    payment_network_address: PaymentNetworkAddress,
+    token_address: TokenAddress,
+) -> Tuple[Optional[PaymentNetworkState], Optional[TokenNetworkState]]:
+    token_network_state = None
+    payment_network_state = chain_state.identifiers_to_paymentnetworks.get(payment_network_address)
+
+    if payment_network_state:
+        token_network_address = payment_network_state.tokenaddresses_to_tokennetworkaddresses.get(
+            token_address
+        )
+
+        if token_network_address:
+            token_network_state = payment_network_state.tokennetworkaddresses_to_tokennetworks.get(
+                token_network_address
+            )
+
+    return payment_network_state, token_network_state
